Completed getting the files in the right place, start of CPU_time_log
I changed the name of simple.nf to CPU_time_log.nf, and made a small start on parsing the smiles to get a IAtomContainer, but this is not finished yet. In the process I accidently made small modification in getSMILES, but this is not anything significant. Short.tsv was still missing from this folder.
git push