Added try/catch statement
The try/catch statement is implemented in order to prevent the process from stopping when a SMILES is missing. The try-catch statement will try the process for every SMILES, and return the catch if the process does not work. Here the catch simply prints all SMILES without a working process.