Write duration per #CPU to tab-delimited tsv file
In order to later be able to plot the duration vs. number of CPUs in Rmarkdown, I made the output of CPU_time_logP.nf a tab-delimited tsv.file containg the number of CPUs and their corresponding calculation durations.  In the previous version I had a for loop to 8 CPUs. However, this starts counting at 0 and therefore didn't work perfectly. I replaced it for a 1.upto(8) loop, to go from 1 to 8 CPUs.