Add tsv file of all canonical and isomeric SMILES
To get this file, I copy the getSMILES.rq file provided by Egon Willighagen to the Wikidata Query Service (https://query.wikidata.org/). The query was run and saved in a tsv file on 24-10-2019. This is the .tsv file uploaded here. The file contains three columns: the wikidata-link, canonical smile (if available) and isomeric smiles (if available) for each compound.