Adapted to use of canonical and isomeric smiles + improved commenting.
The query we use gives us both the canonical and isomeric smiles for each compound (158800 compounds in total). I included an if else statement to use the isomeric smiles for calculating the logP value, but if the isomeric smile is not avalaible, the canonical smile will be used. It is preferred to use the isomeric SMILES because these SMILES take into account the fact that there can be isomeric differences between compounds with the same canonical SMILES. This results in a tab-delimited file containing the number of CPUs and the corresponding calculation times. I also improved the commenting in several places in the code.