Visuallization of results + introduction + methods + discussion
This file contains an introduction, methods, results and discussion on the project. Together with the readme, this is the place where readers can find an overview of all that has been done, and what hasn't been done. This is a html file, to make it easier for readers to see the results, without the need to install anything on their computers.